backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
openai==1.51.0
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
